Do not queue Heartbeat.req messages for longer than 4s
Heartbeat messages are dropped from the request queue if
they are more than 4 seconds old. This is on recommendation
of the OCA to keep the heartbeat interval within 4s of the
interval provided by the CSMS.

If heartbeat messages are queued for longer than 4s,
the resulting interval will exceed this recommendation.